requirements.txt
Flask
Pillow
requests
matplotlib
selenium
gunicorn
